refactor: rename AppMainView.jsx to index.jsx
